ddos/requirements.txt
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
flask>=2.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
imbalanced-learn>=0.11.0
streamlit>=1.25.0
plotly>=5.15.0
PyYAML>=6.0